Add compile support for  MSVC

+++ inc/menu.h
@@ -15,7 +15,7 @@
 //-------------------------------------------------------------------
 
 void show_menu_and_get_settings(Settings *p_s);
-static void draw_menu(Digit ds[], Button bs[], unsigned short num);
-static void menu_main(Digit ds[], Button bs[], unsigned short num);
+static void draw_menu(Digit ds[], Button bs[], unsigned int num);
+static void menu_main(Digit ds[], Button bs[], unsigned int num);
 
 #endif
